alertas en camillas, botiquines y epps al 100% y mas cosas de diseno

+++ inventarios/modulos/camillas/ajax-grid-data.php
@@ -96,7 +96,7 @@ while ($row = mysqli_fetch_array($query)) { // preparing an array
 	$nestedData[] = date("d/m/Y", strtotime($row["FechaRegistro"]));
 	$nestedData[] = '<td><center>
 						<a href="editar.php?CamillaID=' . $row['CamillaID'] . '" id="linkEditar" title="Editar datos" class="btn btn-sm btn-info editar-link"> <i class="fa-solid fa-pen-to-square" style="color: #f2eded;"></i> </a>
-                    	<a href="index.php?action=delete&CamillaID=' . $row['CamillaID'] . '"  data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
+                    	<a href="index.php?action=delete&CamillaID=' . $row['CamillaID'] . '" onClick="return confirmacion();" data-toggle="tooltip" title="Eliminar" class="btn btn-sm btn-danger"> <i class="fa-solid fa-trash-can" style="color: #f2eded;"></i> </a>
                     </center></td>';
 
 	$data[] = $nestedData;
